restructured package to be flatter
and reflective of the current directory structure
